Extract directory traversal partial
